bytebraise-datasmart: overhaul algorithm now that I realize expandKeys() is an essential part.
We don't need to worry about so-called 'synthesized appends/prepends'
because of how expandKeys() and renameVar() works.

It would be interesting however to one day lean more heavily into the
synthesized append/prepends and eliminate the need for expandKeys
altogether. Would need to have get_var consult the list of unexpanded
keys. Not sure what it would buy us.